+++ app/prefs.js
@@ -17,7 +17,6 @@ const default_values = {
     smallscale_guide: false,
     debug: false,
     ignore_hdpi: false,
-    discord: false,
     use_backup: false,
     backup_folder: "",
     fkeys: [ // Stolen mercilously from Pablo, thanks Curtis!